# swagger/openapi.yaml
openapi: 3.0.3

info:
  title: Expense Approval Workflow API
  description: |
    Serverless Intelligent Expense Approval Workflow API.

    Processes employee expense claims through validation, policy checking,
    fraud analysis, and automated decision-making. Supports workflow
    versioning (V1/V2) and manual approval for flagged expenses.

    Key features:
      - Deterministic expense IDs for idempotent submissions
      - Correlation IDs for distributed tracing
      - Paginated employee expense queries via DynamoDB GSI
      - Manual approval endpoint for NEEDS_MANUAL_REVIEW decisions (V2)
  version: 2.0.0

servers:
  - url: http://localhost:5050
    description: Local development server
  - url: https://{apiId}.execute-api.{region}.amazonaws.com/prod
    description: AWS API Gateway (deployed)
    variables:
      apiId:
        default: "your-api-id"
      region:
        default: "us-east-1"

tags:
  - name: Expenses
    description: Expense claim submission and retrieval
  - name: Review
    description: Manual review endpoints
  - name: Health
    description: Service health check
  - name: Development
    description: Local development endpoints (not deployed to AWS)

paths:
  /:
    get:
      summary: Health check
      operationId: healthCheck
      tags: [Health]
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    type: string
                  status:
                    type: string
                  mode:
                    type: string
                  workflowVersion:
                    type: string
                  chaosEnabled:
                    type: boolean
                  timestamp:
                    type: string
                    format: date-time

  /expense:
    post:
      summary: Submit an expense claim
      description: |
        Submit a new expense claim. The expense ID is generated
        deterministically from the content fields, so resubmitting
        the same expense returns 409 with the existing record
        (idempotent). Each request gets a unique correlationId
        for distributed tracing.
      operationId: submitExpense
      tags: [Expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseClaimRequest"
            examples:
              approved:
                summary: Standard meal (APPROVED)
                value:
                  employeeId: "EMP-001"
                  amount: 45.00
                  category: "meals"
                  description: "Team lunch at downtown restaurant"
                  receiptProvided: true
              rejected:
                summary: Over limit (REJECTED)
                value:
                  employeeId: "EMP-002"
                  amount: 500.00
                  category: "meals"
                  description: "Executive dinner at luxury venue"
                  receiptProvided: true
              review:
                summary: Suspicious (NEEDS_MANUAL_REVIEW)
                value:
                  employeeId: "EMP-003"
                  amount: 24.99
                  category: "miscellaneous"
                  description: "test expense"
                  receiptProvided: false
      responses:
        "202":
          description: Expense accepted and processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate submission (idempotent)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  expenseId:
                    type: string
                  correlationId:
                    type: string
                  status:
                    type: string

  /expense/{expenseId}:
    get:
      summary: Get expense by ID
      operationId: getExpense
      tags: [Expenses]
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Expense found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRecord"
        "404":
          description: Not found
    delete:
      summary: Delete expense (local only)
      operationId: deleteExpense
      tags: [Development]
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found

  /expenses/employee/{employeeId}:
    get:
      summary: Get expenses by employee (paginated)
      description: |
        Query all expenses for an employee using the DynamoDB
        EmployeeIndex GSI. Results are paginated. Pass the
        nextToken from the response as a query parameter to
        retrieve the next page.
      operationId: getExpensesByEmployee
      tags: [Expenses]
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
        - name: nextToken
          in: query
          required: false
          description: Pagination token from previous response
          schema:
            type: string
      responses:
        "200":
          description: Paginated expense list
          content:
            application/json:
              schema:
                type: object
                properties:
                  expenses:
                    type: array
                    items:
                      $ref: "#/components/schemas/ExpenseRecord"
                  count:
                    type: integer
                  employeeId:
                    type: string
                  nextToken:
                    type: string
                    nullable: true

  /expenses/{expenseId}/manual-decision:
    post:
      summary: Submit manual review decision (local only)
      description: |
        Approve or reject an expense that was flagged as
        NEEDS_MANUAL_REVIEW. Only valid when the expense
        is in NEEDS_MANUAL_REVIEW or PENDING_REVIEW status.
      operationId: manualDecision
      tags: [Review]
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualDecisionRequest"
            example:
              decision: "APPROVED"
              reason: "Verified with employee, expense is legitimate"
              reviewedBy: "manager@company.com"
      responses:
        "200":
          description: Decision applied
        "400":
          description: Invalid request or expense not reviewable
        "404":
          description: Expense not found

  /expenses:
    get:
      summary: List all expenses (local only)
      operationId: listExpenses
      tags: [Development]
      responses:
        "200":
          description: All expenses
          content:
            application/json:
              schema:
                type: object
                properties:
                  expenses:
                    type: array
                    items:
                      $ref: "#/components/schemas/ExpenseRecord"
                  count:
                    type: integer

components:
  schemas:
    ExpenseClaimRequest:
      type: object
      required: [employeeId, amount, category, description, receiptProvided]
      properties:
        employeeId:
          type: string
          example: "EMP-001"
        amount:
          type: number
          minimum: 0.01
          example: 45.00
        category:
          type: string
          enum: [travel, meals, accommodation, office_supplies, software, training, client_entertainment, transportation, miscellaneous]
        description:
          type: string
          minLength: 3
          description: HTML tags are stripped automatically
        receiptProvided:
          type: boolean

    SubmissionResponse:
      type: object
      properties:
        message:
          type: string
        expenseId:
          type: string
        correlationId:
          type: string
        submittedAt:
          type: string
          format: date-time
        status:
          type: string
        workflowVersion:
          type: string

    ManualDecisionRequest:
      type: object
      required: [decision, reason]
      properties:
        decision:
          type: string
          enum: [APPROVED, REJECTED]
        reason:
          type: string
          minLength: 3
        reviewedBy:
          type: string

    ExpenseRecord:
      type: object
      properties:
        expenseId:
          type: string
        employeeId:
          type: string
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        receiptProvided:
          type: boolean
        submittedAt:
          type: string
          format: date-time
        correlationId:
          type: string
        workflowVersion:
          type: string
        status:
          type: string
          enum: [PROCESSING, APPROVED, REJECTED, NEEDS_MANUAL_REVIEW, PENDING_REVIEW, FAILED_PROCESSING]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string